ai-agent/requirements.txt
openai-agents>=0.8.0
mcp>=1.19.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
redis>=5.0.1
python-dotenv>=1.0.0